Move common part to lib.rs
