+++ src/async.rs
@@ -1,4 +1,13 @@
-use embedded_hal_async::i2c::I2c; // async版のI2cトレイト
+#[cfg(feature = "async")]
+use embedded_graphics_core::{
+    draw_target::DrawTarget,
+    pixelcolor::BinaryColor,
+    Pixel,
+    geometry::Point,
+};
+
+#[cfg(feature = "async")]
+use crate::{cmds::*, BuilderError, Sh1107g, Sh1107gBuilder, DISPLAY_WIDTH, DISPLAY_HEIGHT, BUFFER_SIZE};
 
 #[cfg(feature = "async")]
 impl<I2C, E> Sh1107gBuilder<I2C>
@@ -18,16 +27,6 @@ impl<I2C, E> Sh1107g<I2C>
 where
     I2C: embedded_hal_async::i2c::I2c<Error = E>,
 {
-    // Make new driver instance & Degine function called by the builder
-    // Initialise the internal buffer when called by builder
-    pub fn new(i2c: I2C, address: u8) -> Self {
-        Self {
-            i2c,
-            address,
-            buffer: [0x00; BUFFER_SIZE], // 全てオフで初期化
-        }
-    }
-
     /// Init display
     pub async fn init(&mut self) -> Result<(), E> {
         // 正確な初期化シーケンスの例 (上記のPythonドライバのロジックとデータシートに基づき再構成)
@@ -91,18 +90,13 @@ where
             // 16バイトずつ分割して送信するロジックは理にかなっている。
             // 各ページのデータを16バイトチャンクで送信
             for chunk in page_data.chunks(16) {
-                let mut buf: Vec<u8, 17> = Vec::new(); // 制御バイト1 + データ最大16バイト
+                let mut buf: heapless::Vec<u8, 17> = heapless::Vec::new(); // 制御バイト1 + データ最大16バイト
                 buf.push(0x40).unwrap(); // control byte for data (0x40)
                 buf.extend_from_slice(chunk).unwrap();
                 self.i2c.write(self.address, &buf).await?;
             }
         }
         Ok(())
-    }
-
-    /// 内部バッファをクリアする
-    pub fn clear_buffer(&mut self) {
-        self.buffer.iter_mut().for_each(|b| *b = 0x00);
     }
 }
 
@@ -112,7 +106,7 @@ where
     // I2Cトレイト境界は、DrawTarget自身はI2cトレイトを必要としないため、ここで指定する必要はない
     // むしろ、Sh1107gがI2CとEに依存していることを示すだけでよい
     Sh1107g<I2C>: Sized, // Self::Error が E であることを保証するため
-    E: embedded_hal_async::i2c::Error + embedded_hal::i2c::Error, // 両方のエラー型に対応
+    E: embedded_hal_async::i2c::Error,
 {
     // DrawTarget define color dimension (monochro OLED = BinaryColor)
     type Color = BinaryColor;
